Splitted navigation into different files
