+++ greenhouse-app/app/navigators/components/navigators.d.ts
@@ -1,0 +1,20 @@
+export type NavigatorParamList = {
+  bottomnav: undefined
+  // 🔥 Your screens go here
+  scan: undefined
+  greenhouse: undefined,
+  editPlant: undefined,
+  addPlant: undefined
+}
+
+export type TabParamList = {
+  homepage: undefined
+  // 🔥 Your tabs go here
+  settings: undefined
+}
+
+export type DrawerParamList = {
+  homepage: undefined,
+  profile: undefined,
+  login: undefined
+}
